Refactor hero section layout and styles for improved responsiveness and visual hierarchy
