+++ frontend/components/landing/features/features.module.css
@@ -1,5 +1,5 @@
 .features {
-  @apply relative flex z-20 px-8 lg:px-16 lg:-top-32 lg:-mb-32 max-lg:pt-20 bg-primary;
+  @apply relative flex z-20 px-8 lg:px-16 lg:-top-48 lg:-mb-48 max-lg:pt-20 bg-primary;
 
   & .wave {
     @apply hidden lg:inline absolute left-0 top-0 w-full z-40;
