+++ frontend/lib/constants/landing/interactive-showcase.ts
@@ -3,14 +3,14 @@
     name: 'Galicyjska Kasa Oszczednosci',
     category: 'Historical Landmark',
     schedule: '9.00 a.m. - 10 p.m.',
-    description: 'Description...',
+    description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Assumenda earum molestiae quas quod suscipit velit?',
     image: '/place-carousel.png',
     rating: 4.9,
   },
   {
     name: 'Cozy Café',
     category: 'Food & Drink',
-    description: 'Description...',
+    description: 'At, culpa enim eveniet exercitationem explicabo harum mollitia non numquam quaerat, quidem quo, tempora veritatis.',
     schedule: '9.00 a.m. - 10 p.m.',
     image: '/place-carousel.png',
     rating: 4.9,
@@ -18,7 +18,7 @@
   {
     name: 'Art Gallery',
     category: 'Culture',
-    description: 'Description...',
+    description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Assumenda earum molestiae quas quod suscipit velit?',
     schedule: '9.00 a.m. - 10 p.m.',
     image: '/place-carousel.png',
     rating: 4.9,
@@ -26,7 +26,7 @@
   {
     name: 'Historic Landmark',
     category: 'Sightseeing',
-    description: 'Description...',
+    description: 'At, culpa enim eveniet exercitationem explicabo harum mollitia non numquam quaerat, quidem quo, tempora veritatis.',
     schedule: '9.00 a.m. - 10 p.m.',
     image: '/place-carousel.png',
     rating: 4.9,
